refactor toggles
